requirements.txt
# Core dependencies
click>=8.0.0
PyYAML>=6.0
reportlab>=3.6.0
pandas>=1.5.0

# Development and testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0

# Optional dependencies for future phases
# Uncomment as needed
# pillow>=9.0.0  # For image processing (Phase 8)
# ezdxf>=0.18.0  # For DXF support (Phase 8)